refactor: remove useExpose
